feat: support interpret command

+++ compiler/noirc_evaluator/src/brillig/brillig_ir/codegen_intrinsic.rs
@@ -49,7 +49,7 @@ impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<
 
         // The modulus is guaranteed to fit, since we are truncating down to a bit size that is strictly less than the value_to_truncate.bit_size
         let modulus_var = self.make_constant_instruction(
-            BigInt::from(2_usize).pow(bit_size as u32),
+            BigInt::from(2_usize).pow(bit_size),
             value_to_truncate.bit_size,
         );
